requirements.txt
asgiref==3.3.1
blinker==1.4
Django==3.1.7
django-cors-headers==3.7.0
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.12.2
mongoengine==0.22.1
pymongo==3.11.3
pytz==2021.1
sqlparse==0.4.1
